linuxtage/Reading configuration files with Python/requirements.txt
PyYAML==6.0
pydantic==1.10.7
